change ErrorMessage and how handleError works; add checkIsAuthorized; remove email as key for users but make it unique

+++ src/routes/user.ts
@@ -1,15 +1,14 @@
 import express from "express";
 import {
-  comparePasswords,
   hashPassword,
   handleError,
   getAndThenCacheUser,
   getErrorMessage,
-  getUserOrThrow,
 } from "../helpers";
 import { User } from "../schema/user";
 import { REGISTERED_USERS_CACHE as USERS_CACHE } from "../cache";
-
+import { checkIsAuthorized } from "../middlware/isAuthenticated";
+import { UserDocument } from "../types";
 
 const router = express.Router({
   mergeParams: true,
@@ -21,7 +20,6 @@ router.post("/user", async (req, res) => {
   hashPassword(password, async function (err, hash) {
     try {
       const createdUser = new User({ email, password: hash, hasPaid: false });
-      createdUser._id = email;
       const savedUser = await createdUser.save();
       res.send({
         savedUser,
@@ -42,9 +40,13 @@ router.get("/user/:email", async (req, res) => {
   }
 });
 
-router.delete("/user/:email", async (req, res) => {
-  const { email } = req.params;
+router.delete("/user", async (req, res) => {
+  const { password, email } = req.body;
+
   try {
+    const user = await getAndThenCacheUser(email);
+    console.log({ password, email, user });
+    await checkIsAuthorized(password, user.password);
     const deletedUser = await User.deleteOne({ email });
     if (deletedUser.deletedCount > 0) {
       USERS_CACHE.delete(email);
@@ -56,45 +58,47 @@ router.delete("/user/:email", async (req, res) => {
   }
 });
 
-// router.put("/user/:currentPasswordHash", async (req, res) => {
-//   const userToUpdate = req.body;
-//   const currentPassword = req.params;
-//   const { email, password, hasPaid } = req.body;
+router.put("/user", async (req, res) => {
+  const userToUpdate = req.body as UserDocument;
+  const { _id, email, password, hasPaid } = req.body;
 
-//   try {
-//     const user = await getAndThenCacheUser(email);
-
-//     comparePasswords(currentPassword, user.password);
-//     hashPassword(password, async (err, hash) => {
-//       try {
-//         if (err || !hash) {
-//           res
-//             .status(500)
-//             .send(getErrorMessage(`Unable to update user with '${email}'.`));
-//         }
-//         const updatedUser = await User.updateOne({ email }, userToUpdate);
-//         console.log({ updatedUser });
-//         if (updatedUser.modifiedCount > 1) {
-//           USERS_CACHE.delete(email);
-//         }
-//         res.send(updatedUser);
-//       } catch (error) {
-//         handleError(res, error);
-//       }
-//     });
-//   } catch (error) {
-//     handleError(res, error);
-//   }
-// });
-
-router.post("/user/isPasswordSame", async (req, res) => {
-  const { email, password } = req.body;
   try {
-    const user = await getUserOrThrow(email);
-    comparePasswords(password, user.password, function (err, isSame) {
-      res.send(isSame);
-      return;
+    const user = await getAndThenCacheUser(email);
+    const isPasswordSame = await checkIsAuthorized(password, user.password);
+    console.log({ isPasswordSame });
+    hashPassword(password, async (err, hash) => {
+      try {
+        if (err || !hash) {
+          res
+            .status(500)
+            .send(getErrorMessage(`Unable to update user with '${email}'.`));
+        }
+        const updatedUser = await User.updateOne(
+          { email },
+          {
+            ...userToUpdate,
+            password: isPasswordSame ? userToUpdate.password : hash,
+          }
+        );
+        console.log({ updatedUser });
+        if (updatedUser.modifiedCount > 1) {
+          USERS_CACHE.delete(email);
+        }
+        res.send(updatedUser);
+      } catch (error) {
+        handleError(res, error);
+      }
     });
+  } catch (error) {
+    handleError(res, error);
+  }
+});
+
+router.post("/user/isEmailAvailable", async (req, res) => {
+  const { email } = req.body;
+  try {
+    const user = await User.findOne({ email });
+    res.send(user?.email == email);
   } catch (error) {
     handleError(res, error);
   }
